refactor: Move `FirstWhereOrNull` into an util extension file and update imports
