+++ lib/core/electrum/data/repository/electrum_server_repository_impl.dart
@@ -145,12 +145,3 @@ class ElectrumServerRepository {
     return server;
   }
 }
-
-extension FirstWhereOrNullExtension<T> on Iterable<T> {
-  T? firstWhereOrNull(bool Function(T element) test) {
-    for (final element in this) {
-      if (test(element)) return element;
-    }
-    return null;
-  }
-}
